Interface-ified TextureSet
