+++ src/SA3D.Texturing/TextureFileUtilities.cs
@@ -16,6 +16,7 @@ using SixLabors.ImageSharp.Formats.Tiff;
 using SixLabors.ImageSharp.Formats.Webp;
 using SixLabors.ImageSharp.PixelFormats;
 using System;
+using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.IO;
 
@@ -573,6 +574,110 @@ namespace SA3D.Texturing
 				Name = filename,
 				IsIndex4 = isIndex4
 			};
+		}
+
+
+		/// <summary>
+		/// Writes a content index used by texture packs to a writer.
+		/// </summary>
+		/// <param name="textureSet">The texture set for which to write the content index</param>
+		/// <param name="writer">The writer to write to.</param>
+		/// <param name="nameSuffix">Suffix for every texture name.</param>
+		public static void WriteContentIndex(this ITextureSet textureSet, TextWriter writer, string nameSuffix)
+		{
+			foreach(ITexture texture in textureSet.Textures)
+			{
+				writer.WriteLine($"{texture.GlobalIndex},{texture.Name}{nameSuffix},{texture.OverrideWidth}x{texture.OverrideHeight}");
+			}
+		}
+
+		/// <summary>
+		/// Generates a content index used by texture packs.
+		/// </summary>
+		/// <param name="textureSet">The texture set for which to write the content index</param>
+		/// <param name="nameSuffix">Suffix for every texture name.</param>
+		/// <returns>The index contents.</returns>
+		public static string WriteContentIndexToString(this ITextureSet textureSet, string nameSuffix)
+		{
+			using StringWriter writer = new();
+			WriteContentIndex(textureSet, writer, nameSuffix);
+			return writer.ToString();
+		}
+
+		/// <summary>
+		/// Writes a content index used by texture packs to a file.
+		/// </summary>
+		/// <param name="textureSet">The texture set for which to write the content index</param>
+		/// <param name="filepath">Path of the file to write to.</param>
+		/// <param name="nameSuffix">Suffix for every texture name.</param>
+		/// <returns>The index contents.</returns>
+		public static void WriteContentIndexToFile(this ITextureSet textureSet, string filepath, string nameSuffix)
+		{
+			using StreamWriter writer = File.CreateText(filepath);
+			WriteContentIndex(textureSet, writer, nameSuffix);
+		}
+
+
+		/// <summary>
+		/// Exports the texture set as a texture pack useable by sonic adventure modloaders.
+		/// </summary>
+		/// <param name="textureSet">The texture set to export</param>
+		/// <param name="outDirectory">The directory to which to write the files.</param>
+		/// <param name="format">Format to write images in</param>
+		public static void ExportTexturePack(this ITextureSet textureSet, string outDirectory, ImageFormat format)
+		{
+			string extension = "." + format.ToString().ToLowerInvariant();
+			string indexPath = Path.Join(outDirectory, "index.txt");
+			WriteContentIndexToFile(textureSet, indexPath, extension);
+
+			foreach(ITexture texture in textureSet.Textures)
+			{
+				string path = Path.Join(outDirectory, texture.Name + extension);
+
+				if(texture is IndexTexture indexTex)
+				{
+
+					indexTex.WriteIndexImageToFile(path, format, false);
+				}
+				else
+				{
+					texture.WriteImageToFile(path, format);
+				}
+			}
+		}
+
+		/// <summary>
+		/// Imports texture from a texture pack useable by sonic adventure modloaders.
+		/// </summary>
+		/// <param name="directory">The directory from which to read the files.</param>
+		/// <returns>The imported texture set.</returns>
+		public static TextureSet ImportTexturePack(string directory)
+		{
+			List<ITexture> textures = [];
+
+			string indexPath = Path.Join(directory, "index.txt");
+			string[] index = File.ReadAllLines(indexPath);
+
+			foreach(string item in index)
+			{
+				string[] values = item.Split(',');
+				string filename = values[1];
+
+				string texturePath = Path.Join(directory, filename);
+				Texture texture = TextureFileUtilities.ReadImageFromFile(texturePath);
+
+				texture.GlobalIndex = uint.Parse(values[0]);
+				if(values.Length >= 3)
+				{
+					string[] overrideDimensions = values[2].Split('x');
+					texture.OverrideWidth = int.Parse(overrideDimensions[0]);
+					texture.OverrideHeight = int.Parse(overrideDimensions[1]);
+				}
+
+				textures.Add(texture);
+			}
+
+			return new(textures.ToArray());
 		}
 	}
 }
